requirements.txt
config-parser==0.0.1
crc==5.0.0
pyusb==1.2.1
